Fix four defects the scanner review found
An adversarial review over five dimensions, each finding put through two
independent refutation attempts with both required to agree. Four real
defects, one of which is worse than anything I had spotted myself.

CRITICAL - two unidentified devices merged into one.

DeviceRecord::serial_number was an int, and to_record cast the wire's
uint32_t into it. The all-FF uninitialised value 0xFFFFFFFF became -1.
has_stable_identity() tested `serial_number != 0`, and -1 is not 0, so it
called that a real identity - and add_or_merge merged every all-FF device
into a single record. unidentified_count() then reported none of them.

A technician scanning a site would have been told there was one nameless
device when there were several. That is precisely the collapse the
"unidentified devices never merge" test exists to prevent; the test
passed because it only ever used serial 0.

The field is uint32_t now and has_stable_identity defers to
is_uninitialised_serial, so both values are covered. Confirmed the
regression test bites by restoring the old predicate: three all-FF
devices collapse to 1 and unidentified_count returns 0.

HIGH - an approved repair outlived the problem it described.

Clearing approvals sat inside the branch that replaced the repair list,
so a rescan finding NO problems left the old repair in place and still
approved - approved, and describing a device state that no longer
existed. Approval is now withdrawn on every merge unconditionally.
Re-approving costs a click; holding consent given under different
information does not have a bounded cost.

That needed a way to tell a complete look at a device from a partial one,
since a full scan finding nothing must clear repairs while a serial sweep
that learned only an address must not. DeviceRecord::observation_complete
marks the difference and to_record sets it.

HIGH - the scan loop could run forever.

Its elapsed counter only advanced when receive() TIMED OUT, so any
datagram arriving before the deadline advanced it by nothing. Continuous
broadcast traffic - from anything on the subnet, not necessarily our
devices - would keep the loop alive indefinitely. The device cap does not
save it, because ignored and malformed datagrams produce no device. Now
measured against steady_clock, with a datagram cap as a second bound so a
flood cannot spin it hot.

I had found this one independently before the review reported it.

MEDIUM - a test that could not fail.

test_every_field_round_trips claims to check every field but omitted
subnet_protocol, and the packet set it to 0 while the struct defaults to
0 - so the assertion would have passed even if the parser never read the
field. Set to 12 on the wire now and asserted.

One finding was correctly killed, and checking it myself confirms the
kill: the object_instance byte order looks wrong but matches T3000
exactly. The wire order is deliberately scrambled - offset 34 holds
instance_2 and 35 holds instance_1 - and TStatScanner.cpp:2286 assembles
by field name, not by position. The reviewer assumed wire order should
map to bit order.

704 checks, 0 failures.

+++ T5000/discovery/scanner_selftest.cpp
@@ -117,7 +117,7 @@ namespace
         check_eq((int)result.devices.size(), 1, "one device");
 
         const auto& d = result.devices[0];
-        check_eq(d.serial_number, 500123, "serial");
+        check_eq((long)d.serial_number, 500123, "serial");
         check(d.product == ProductClassId::Esp32T3Series, "product id mapped");
         check_eq(d.firmware, 538, "firmware, which the PTP gate needs");
         check(d.provenance == Provenance::BacnetBroadcast, "provenance");
